- Resolved SensorData.cs: use int SoilMoisture (matches server-side calc), keep RawMoisture
- Resolved SensorService.cs: keep dev's CalculateSoilMoisture/Light/WindSpeed, add RawMoisture
  persistence, fix ESP32 12-bit ADC validation cap (1023 → 4095)
- Resolved sketch_may10a.ino: combine SHTC3 temp/humidity (ours) + APDS9250 light sensor
  (dev) + wind pins GPIO 32/33 (dev) + BLE broadcast (ours); HTTP JSON includes all raw
  fields for server-side calculation
- Added APDS9250 library to platformio.ini

+++ backend/Goalz/Goalz.Domain/Entities/SensorData.cs
@@ -7,7 +7,7 @@ public class SensorData
     public long Light { get; set; }
     public long Humidity { get; set; }
     public double Temp { get; set; }
-    public long SoilMoisture { get; set; }
+    public int SoilMoisture { get; set; }
     public double Wind { get; set; }
     public int RawMoisture { get; set; }
     public DateTime Timestamp { get; set; }
